refactor: add rules as a module
