+++ src/modules/global-messages/GlobalMessage.tsx
@@ -7,7 +7,7 @@ import {GlobalMessageContextEnum, GlobalMessageType} from './types';
 import {getTextForLanguage} from '@atb/translations';
 import {useNow} from '@atb/utils/use-now';
 import {isWithinTimeRange} from '@atb/utils/is-within-time-range';
-import {RuleVariables} from '@atb/rule-engine/rules';
+import {RuleVariables} from '@atb/modules/rule-engine';
 import {ContrastColor, TextColor} from '@atb/theme/colors';
 import {MessageInfoText} from '@atb/components/message-info-text';
 import {useAuthContext} from '@atb/auth';
