+++ src/modules/global-messages/index.ts
@@ -4,5 +4,5 @@ export {
 } from './GlobalMessagesContext';
 
 export {GlobalMessage} from './GlobalMessage';
-export type {GlobalMessageType} from './types';
+export type {GlobalMessageType, AppPlatformType} from './types';
 export {GlobalMessageContextEnum} from './types';
